+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 All Notable changes to `flipboxdigital\interval` will be documented in this file
 
-## Unreleased
+## 1.0.0-beta.4 - 2017-08-30
 ### Fixed
 - Issue where human duration string would not return negative if DateInterval is inverted